functions/flickr/requirements.txt
python-telegram-bot
flickrapi